refactor(main): extract FileLicenseStorage and window state to services
First scoped slice of the PR-G "split god files" effort.

apps/desktop/src/main/index.ts goes from 1065 -> 950 lines (-11%).
No behavior change. All extractions are self-contained and surface-area
preserving:

- services/fileLicenseStorage.ts — was embedded in index.ts at lines
  146-216. The three readJsonOrNull helpers fold the repeated
  try/catch/return-null pattern into one private function. The TODO
  about signed subscription envelopes points back at PR-F3 so the
  next step is unambiguous.

- services/windowState.ts — was at lines 218-254 of index.ts.
  Comment now explicitly documents why the file I/O is sync (called
  during BrowserWindow construction before the renderer mounts and
  during close where handlers don't await).

What this PR DELIBERATELY does NOT touch:

- packages/storage-sqlite/src/repositories/SQLiteNoteRepository.ts
  (1121 lines). Splitting this into NoteCrudRepository +
  NoteTagRepository + NoteArchiveRepository + NoteSyncRepository
  needs runtime verification against a real DB and risks merge
  conflicts with every other PR in flight. Separate effort.
- apps/desktop/src/renderer/components/MarkdownEditor.tsx (724
  lines). The CodeMirror surface area is too entangled to split
  without an E2E suite to catch regressions. Defer until PR-E
  (#277) is stabilized and we can refactor under test coverage.

Validates:
- pnpm -r typecheck — green
- pnpm test — 17/17 packages
- pnpm --filter @readied/desktop typecheck — green (includes e2e
  tsconfig)
